DemoInfer

Documentation for DemoInfer.

Module to run demographic inference on diploid genomes, under the assumption of panmixia (i.e. the inferred effective population size is half the inverse of the observed mean coalescence rate). See this repo for a demo of how to use it.

Data form of input and output

The genome needs to be SNP-called and the genomic distance between consecutive heterozygous sites needs to be computed. The input is then a vector containind such distances. Additionally, mutation and recombination rates need to be chosen and passed as input as well.

The demographic model underlying the inference is composed of a variable number of epochs and the population size is constant along each epoch.

The output is a vector of parameters in the form [L, N0, T1, N1, T2, N2, ...] where L is the genome length, N0 is the ancestral population size in the infinite past and the subsequent pairs $(T_i, N_i)$ are the duration and size of following epochs going from past to present. This format is referred to as TN vector throughout.

DemoInfer.FitResult

DemoInfer.compare_models

DemoInfer.durations

DemoInfer.estimate_nepochs

DemoInfer.evd

DemoInfer.fit

DemoInfer.get_chain

DemoInfer.get_para

DemoInfer.get_sim!

DemoInfer.pop_sizes

DemoInfer.pre_fit

DemoInfer.sds

DemoInfer.FitResult — Type

struct FitResult

A data structure to store the results of a fit.

Some methods are defined for this type to get the vector of parameters, std errors, model evidence, etc. See get_para, sds, evd, pop_sizes, durations.

DemoInfer.compare_models — Method

compare_models(models::Vector{FitResult})

Compare the models parameterized by FitResults and return the best one.

Theoretical explanation

TBD

DemoInfer.durations — Method

durations(fit::FitResult)

Return the fitted durations of the epochs.

DemoInfer.estimate_nepochs — Method

estimate_nepochs(h::Histogram, mu::Float64, Ltot::Number; max_nepochs::Int = 10, kwargs...)

Estimate the number of epochs needed to fit the histogram h.

The mutation rate mu is assumed to be per base pairs per generation, and the total length of the genome Ltot is in base pairs.

The optional argument max_nepochs defines the maximum number of epochs that are explored, while the other keyword arguments are passed to pre_fit.

DemoInfer.evd — Method

evd(fit::FitResult)

Return the evidence of the fit.

DemoInfer.fit — Method

fit(h_obs::Histogram, nepochs::Int, mu::Float64, rho::Float64, Ltot::Number; kwargs...)
fit(h_obs, nepochs, mu, rho, Ltot, init::Vector{Number}; kwargs...)

Fit iteratively h_obs with a demographic history of piece-wise constant nepochs.

Return a vector of FitResult, see FitResult, mu and rho are the mutation and recombination rates, respectively, per base pair per generation and Ltot is the total length of the genome, in base pairs. Optional argument init can be used to provide an initial point for the iterations.

Arguments

iters::Int=100: The number of iterations to perform. Suggested value is at least 10.

burnin::Int=1: The number of iterations to discard as burnin. Notice that in general already the second iteration is a good sample.

allow_boundary::Bool=false: If true, the function will allow the fit to reach the upper

boundaries of populations sizes. This can be useful if structure is expected because epochs of separation between two subpopulations show up as a higher population size (smaller coalescence rate).

level::Float64=0.95: The confidence level for the confidence intervals on the parameters estimates.

Tlow::Int=10: The lower bound for the duration of epochs.

Nlow::Int=10, Nupp::Int=100000: The lower and upper bounds for the population sizes.

smallest_segment::Int=30: The smallest segment size to consider for the optimization.

This dictates the lower bound which is considered in when initially guessing when to insert a new epoch, i.e. the upper bound for time splits. Notice that the optimization per se is not affected by this limit.

DemoInfer.get_chain — Method

get_chain(fit::FitResult)

Return two matrices containing the chain of fitted parameters and std errors respectively.

DemoInfer.get_para — Method

pars(fit::FitResult)

Return the parameters of the fit.

DemoInfer.get_sim! — Method

get_sim!(params::Vector, h::Histogram, mu::Float64, rho::Float64; factor = 1)

Simulate a population according to the demographic parameters in params and stores the IBS segments in the h.

Arguments

'params': vector of the form [L, N0, T1, N1, T2, N2] where L is the genome length, N0 is the ancestral population size and successive pairs of Ti and Ni are the durations and sizes of subsequent epochs.

factor: determine how many genomes are simulated and averaged

DemoInfer.pop_sizes — Method

pop_sizes(fit::FitResult)

Return the fitted population sizes, from past to present.

DemoInfer.pre_fit — Method

pre_fit(h::Histogram, nfits::Int, mu::Float64, Ltot::Number; kwargs...)

Preliminarily fit h with an approximate model of piece-wise constant nepochs.

The mutation rate mu is assumed to be per base pairs per generation and the total length of the genome Ltot is in base pairs. The fit approximate the histogram with more and more epochs up to nepochs, so the result is a vector of FitResult, one for each number of epochs.

See also FitResult.

Arguments

Tlow::Int=10: The lower bound for the duration of epochs.

Nlow::Int=10, Nupp::Int=100000: The lower and upper bounds for the population sizes.

smallest_segment::Int=30: The smallest segment size to consider for the optimization,

same as in fit.

DemoInfer.sds — Method

sds(fit::FitResult)

Return the standard deviations of the parameters of the fit.