DemoInfer

Documentation for DemoInfer.

DemoInfer.corrected_fit

DemoInfer.get_evidence

DemoInfer.get_sds

DemoInfer.get_sim!

DemoInfer.sequential_fit

DemoInfer.corrected_fit — Method

corrected_fit(h_obs::Histogram, nepochs::Int, mu::Float64, rho::Float64; kwargs...)

Fit iteratively the histogram h_obs with up to nepochs and return a vector of FitResult and a vector of matrices chain with the parameter vector of each iteration.

Arguments

h_obs::Histogram: The histogram to fit.

nepochs::Int: The maximum number of epochs to fit.

mu::Float64: The mutation rate.

rho::Float64: The recombination rate.

Ltot::Number: The total length of the genome.

start::Int=1: The optional starting epoch.

iters::Int=100: The number of iterations to perform. Suggested value is at least 10.

final_factor::Int=100: The factor for the final simulation (i.e. how many genomes are simulated).

If fitting a cumulative histogram it should be 1.

allow_boundary::Bool=false: If true, the function will allow the fit to reach the upper

boundaries of populations sizes. This can be useful if structure is expected because epochs of separation between two subpopulations show up as a higher population size (smaller coalescence rate).

level::Float64=0.95: The confidence level for the confidence intervals on the parameters estimates.

DemoInfer.get_evidence — Method

get_evidence(fit::FitResult)

Return the evidence of the fit.

DemoInfer.get_sds — Method

get_sds(fit::FitResult)

Return the standard deviations of the parameters of the fit.

DemoInfer.get_sim! — Method

get_sim!(params::Vector, h::Histogram, mu::Float64, rho::Float64; factor = 1)

Simulate a population according to the demographic parameters in params and stores the IBS segments in the h.

Arguments

'params': vector of the form [L, N0, T1, N1, T2, N2] where L is the genome length, N0 is the ancestral population size and successive pairs of Ti and Ni are the durations and sizes of subsequent epochs.

factor: determine how many genomes are simulated and averaged

DemoInfer.sequential_fit — Method

sequential_fit(h::Histogram, μ::Float64, nfits::Int; Tlow = 10)

Fit (uncorrected) the histogram h with an increasing number of epochs, starting from 1 up to nfits.

The function returns an array of FitResult objects, one for each fit. Tlow is the optional lower bound for the duration of epochs.